m12n: Move OakVersion to the top-level org.apache.jackrabbit.oak package
This class is not (yet) used outside of oak-core and is an internal
detail of Oak.
