+++ oak-core/src/main/java/org/apache/jackrabbit/oak/OakVersion.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.oak.util;
+package org.apache.jackrabbit.oak;
 
 import java.io.IOException;
 import java.io.InputStream;
